requirements.txt
aiogram>=3.15
aiohttp>=3.10
aiosqlite>=0.20
python-dotenv>=1.0
